Move resource parser method from SN resource manager to Services to remove System.Web.Compilation dependency.
