+++ src/Services/SenseNet.Services.csproj
@@ -294,6 +294,7 @@
     <Compile Include="PortalSettings.cs" />
     <Compile Include="Preview\WatermarkServicesTemplateReplacer.cs" />
     <Compile Include="Security\UserAccessProvider.cs" />
+    <Compile Include="ServiceExtensions.cs" />
     <Compile Include="ServiceTools.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
     <Compile Include="QueryBuilder.cs" />
